tools/t1/requirements.txt
numpy
opencv-python
scikit-image
Pillow
pytest
PySide6
pytest-qt
